+++ src/constants/articleProps.ts
@@ -177,3 +177,13 @@ export const defaultArticleState = {
 };
 
 export type ArticleStateType = typeof defaultArticleState;
+
+export type StylesToChange = {
+	'--font-family': string,
+	'--font-size': string,
+	'--font-color': string,
+	'--container-width': string,
+	'--bg-color': string
+}
+
+export type OnSubmitChange = (newStyles: StylesToChange) => void;
